fix: improving person helped screens
