+++ src/services/Api.tsx
@@ -1,5 +1,5 @@
 import axios from 'axios';
 
-const api = axios.create({ baseURL: 'http://localhost:8082/api' });
+const api = axios.create({ baseURL: 'http://localhost:8000/api' });
 
 export default api;
